add odd-even game
